only use one public create_tantivy_doc() method
