+++ src/rest_api/api/indexing/bulk_index.rs
@@ -1,18 +1,16 @@
-use crate::indexing::{combine_json_objects, index_manager::IndexManager, KNOWN_DESCRIPTORS};
+use crate::indexing::{create_tantivy_doc, index_manager::IndexManager, KNOWN_DESCRIPTORS};
 use crate::rest_api::api::{
     BulkRequest, PostIndexBulkResponseError, PostIndexBulkResponseOk,
     PostIndexBulkResponseOkStatus, PostIndexesBulkIndexResponse,
 };
 use crate::search::compound_processing::process_cpd;
-use crate::search::scaffold_search::{scaffold_search, PARSED_SCAFFOLDS};
 use crate::search::similarity_search::encode_fingerprints;
 use poem_openapi::payload::Json;
 use rayon::prelude::*;
-use serde_json::Value;
 use std::collections::HashMap;
 use bitvec::prelude::BitVec;
-use rdkit::{Fingerprint, ROMol};
-use tantivy::{doc, Document};
+use rdkit::Fingerprint;
+use tantivy::Document;
 use tantivy::schema::{Field, Schema};
 
 pub async fn v1_post_index_bulk(
@@ -166,56 +164,4 @@ fn batch_doc_creation(
         }).collect::<Vec<_>>();
 
     Ok(docs)
-}
-
-fn create_tantivy_doc(
-    canon_taut: &ROMol,
-    extra_data: &Option<Value>,
-    pattern_fp: &Fingerprint,
-    morgan_fp: &Fingerprint,
-    descriptors: &HashMap<String, f64>,
-    similarity_cluster: i32,
-    smiles_field: Field,
-    pattern_fingerprint_field: Field,
-    morgan_fingerprint_field: Field,
-    descriptor_fields: &HashMap<&str, Field>,
-    extra_data_field: Field,
-    other_descriptors_field: Field,
-) -> eyre::Result<impl Document> {
-    let mut doc = doc!(
-        smiles_field => canon_taut.as_smiles(),
-        pattern_fingerprint_field => pattern_fp.0.as_raw_slice(),
-        morgan_fingerprint_field => morgan_fp.0.as_raw_slice(),
-    );
-
-    let scaffold_matches = scaffold_search(&pattern_fp.0, &canon_taut, &PARSED_SCAFFOLDS)?;
-    let scaffold_json = match scaffold_matches.is_empty() {
-        true => serde_json::json!({"scaffolds": vec![-1]}),
-        false => serde_json::json!({"scaffolds": scaffold_matches}),
-    };
-
-    let cluster_json = serde_json::json!({"similarity_cluster": similarity_cluster});
-
-    let other_descriptors_json = combine_json_objects(Some(scaffold_json), Some(cluster_json));
-
-    if let Some(other_descriptors_json) = other_descriptors_json {
-        doc.add_field_value(other_descriptors_field, other_descriptors_json);
-    }
-
-    if let Some(extra_data) = extra_data {
-        doc.add_field_value(extra_data_field, extra_data.clone());
-    }
-
-    for field in KNOWN_DESCRIPTORS {
-        if let Some(val) = descriptors.get(field) {
-            if field.starts_with("Num") || field.starts_with("lipinski") {
-                let int = *val as i64;
-                doc.add_field_value(*descriptor_fields.get(field).unwrap(), int);
-            } else {
-                doc.add_field_value(*descriptor_fields.get(field).unwrap(), *val);
-            };
-        }
-    }
-
-    Ok(doc)
 }
